bot/requirements.txt
# Django Core
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9 
django-phonenumber-field==7.3.0
phonenumbers==8.13.27
django-filter==23.5
drf-spectacular==0.27.0
cloudinary==1.41.0
django-cloudinary-storage==0.3.0
python-decouple==3.8
django-redis==5.4.0
redis==5.0.1
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
celery==5.3.4
twilio==8.11.1 
requests==2.31.0 
boto3==1.34.19
django-storages==1.14.2
Pillow==10.2.0
pytz==2023.3.post1
django-debug-toolbar==4.2.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.39.2
python-telegram-bot>=20.4
asgiref>=3.7
